REFACTOR :: features, hook
