+++ src/app/notice/[detail]/page.tsx
